moved model tests over to rawlsModel
